Rename chunker/ -> chunkers/ for consistency with parsers/
